refactored profile-jobs-list-message-screen
