+++ lib/src/common/common_exports.dart
@@ -3,3 +3,4 @@ export 'ea_kazi_scaffold.dart';
 export 'search_field.dart';
 export 'reusable_card_widget.dart';
 export 'ea_kazi_app_bar.dart';
+export 'custom_divider.dart';
