+++ lib/src/constants/app_sizes.dart
@@ -16,6 +16,7 @@ class Sizes {
   static const p20 = 20.0;
   static const p24 = 24.0;
   static const p32 = 32.0;
+  static const p40 = 40.0;
   static const p48 = 48.0;
   static const p64 = 64.0;
   static const p80 = 80.0;
