+++ lib/src/features/authentication/authentication_exports.dart
@@ -1,3 +1,4 @@
 export 'presentation/onboarding/onboarding_screen.dart';
 export 'presentation/sign_in/sign_in_screen.dart';
 export 'presentation/sign_in/sign_up_screen.dart';
+export 'presentation/profile/profile_screen.dart';
