+++ lib/src/features/jobs&courses/presentation/courses/course_detail_screen.dart
@@ -11,7 +11,6 @@ class CourseDetailScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    SizeConfig().init(context);
     return Scaffold(
       appBar: PreferredSize(
         preferredSize: Size.fromHeight(AppBar().preferredSize.height),
