+++ lib/src/features/jobs&courses/presentation/courses/course_video_screen.dart
@@ -14,7 +14,6 @@ class CourseVideoScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    SizeConfig().init(context); //TODO: REMOVE THIS LINE
     return Scaffold(
       appBar: PreferredSize(
         preferredSize: Size.fromHeight(AppBar().preferredSize.height),
