+++ lib/src/features/jobs&courses/presentation/home/home_components/recommendation_widget.dart
@@ -1,4 +1,7 @@
+import 'package:auto_route/auto_route.dart';
 import 'package:ea_kazi/src/constants/constants_exports.dart';
+import 'package:ea_kazi/src/features/jobs&courses/presentation/jobs/jobs_screen.dart';
+import 'package:ea_kazi/src/routes/router.gr.dart';
 import 'package:flutter/material.dart';
 
 class RecommendationWidget extends StatelessWidget {
@@ -18,7 +21,7 @@ class RecommendationWidget extends StatelessWidget {
             style: Theme.of(context).textTheme.subtitle1,
           ),
           TextButton(
-            onPressed: () => null,
+            onPressed: () => context.router.push(const JobsRoute()),
             //TODO: IMPLEMENT SEE ALL RECOMMENDATIONS
             child: Text(
               AppStrings.seeAll,
